requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
Pillow==10.4.0
gunicorn==21.2.0
